Add 'fail_on_commit' option to DevelopmentBranchCheckTask
This change introduces a new 'fail_on_commit' option to the DevelopmentBranchCheckTask. This option, when set to false, allows tasks to fail in a non-blocking manner during a git pre-commit, while still failing as expected during a run context. The default value is set to true, maintaining the current behavior unless overridden.